requirements.txt
google-generativeai==0.8.3
faiss-cpu==1.9.0
scikit-learn==1.6.1
xgboost==2.1.3
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
PyMuPDF==1.25.1
python-docx==1.1.2
Jinja2==3.1.5
pandas==2.2.3
numpy==2.2.1
joblib==1.4.2
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3
